requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.0
